fix: TM-452 add new personal hook for the dates and deleted comments
