+++ src/admin/components/ResourceTabs/InformationTab/components/InformationTabRow.tsx
@@ -4,34 +4,11 @@ import { Else, If, Then, When } from "react-if";
 
 import List from "@/components/extensive/List/List";
 import { FormSummaryRowProps, useGetFormEntries } from "@/components/extensive/WizardForm/FormSummaryRow";
+import { useVerifyDate } from "@/hooks/useVerifyDate";
 
 const InformationTabRow = ({ index, ...props }: FormSummaryRowProps) => {
   const entries = useGetFormEntries(props);
-
-  const formatDateString = (inputDateString: string) => {
-    if (!/^\d{4}-\d{2}-\d{2}$/.test(inputDateString)) {
-      return "Invalid Date";
-    }
-
-    const [year, month, day] = inputDateString.split("-");
-    const dateObject = new Date(Number(year), Number(month) - 1, Number(day));
-
-    const formattedDay = dateObject.getDate().toString().padStart(2, "0");
-    const formattedMonth = (dateObject.getMonth() + 1).toString().padStart(2, "0");
-    const formattedYear = dateObject.getFullYear();
-
-    const formattedDate = `${formattedDay}/${formattedMonth}/${formattedYear}`;
-
-    return formattedDate;
-  };
-
-  function isDateStringValid(dateString: string): boolean {
-    if (!isNaN(Number(dateString))) {
-      return false;
-    }
-    const timestamp = Date.parse(dateString);
-    return !isNaN(timestamp);
-  }
+  const { formatDateString, isDateStringValid } = useVerifyDate();
 
   return (
     <>
