+++ src/admin/modules/organisations/components/OrganisationsList.tsx
@@ -78,7 +78,6 @@ const ApplicationDataGrid = () => {
       </Tabs>
       <Divider />
       <Datagrid>
-        {/* format change Organisation TM-452 */}
         <DateField source="created_at" label="Date Created" locales="en-GB" />
         <TextField source="name" label="Organisation Name" />
         <SelectField label="Organisation Type" source="type" choices={optionToChoices(getOrganisationTypeOptions())} />
